generate choice situations from actual poker data
